gateway/pyproject.toml
[project]
name = "opsgenie-gateway"
version = "0.1.0"
description = "Gateway service for OpsGenie-for-Agents"
authors = [{ name = "OpsGenie-for-Agents" }]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "SQLAlchemy>=2.0,<3.0",
    "psycopg[binary]>=3.1,<4.0",
    "python-dotenv>=1.0,<2.0",
    "pydantic>=2.8,<3.0",
    "pydantic-settings>=2.4,<3.0",
    "jsonschema>=4.21,<5.0",
    "prometheus-client>=0.20,<1.0",
    "opentelemetry-api>=1.26,<2.0",
    "opentelemetry-sdk>=1.26,<2.0",
    "opentelemetry-exporter-otlp>=1.26,<2.0",
    "opentelemetry-instrumentation-fastapi>=0.47b0,<1.0",
    "opentelemetry-instrumentation-logging>=0.47b0,<1.0",
    "opentelemetry-instrumentation-requests>=0.47b0,<1.0",
    "httpx>=0.27,<1.0",
    "pyyaml>=6.0,<7.0",
    "sse-starlette>=1.8,<2.0",
    "temporalio>=1.7",
    "openai>=1.0,<2.0",
    "anthropic>=0.25,<1.0",
    "passlib[bcrypt]>=1.7.4,<2.0",
    "kubernetes>=29.0.0",
    "authlib>=1.3,<2.0",
    "itsdangerous>=2.2,<3.0",
    "PyJWT>=2.8,<3.0",
    "stripe>=7.0,<8.0",
    "typing-extensions>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0",
    "ruff>=0.6.0",
    "pytest>=8.0",
    "httpx>=0.27",
    "pytest-asyncio>=0.23",
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",
    "F",
    "I",
    "B",
    "UP",
]
ignore = []

[tool.pytest.ini_options]
testpaths = ["gateway/tests"]
addopts = "-q"

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"
